requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.1
Pillow==9.2.0
requests==2.25.1
